Ændret model klasse og tilføjet item som List samt en item klasse. QODANA TEST
