--- src/main/java/com/gruppe8/wishlist/model/WishlistModel.java
@@ -1,38 +1,0 @@
-package com.gruppe8.wishlist.model;
-
-public class WishlistModel {
-
-    private String name;
-    private String description;
-    private int price;
-
-public WishlistModel(String name, String description, int price){
-    this.name = name;
-    this.description = description;
-    this.price = price;
-}
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public void setPrice(int price) {
-        this.price = price;
-    }
-}
